# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agustinos.tv API",
        "version": "3.20.01"
    },
    "servers": [
        {
            "url": "http://api.agustinos.tv"
        }
    ],
    "paths": {
...
        "/api/eventos": {
            "get": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Listar eventos",
                "description": "Obtener la lista de eventos",
                "operationId": "App\\Http\\Controllers\\EventosController::index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de Página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "canal_id",
                        "in": "query",
                        "description": "Id de el canal",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista resultado de los eventos"
                    }
                }
            }
        },
        "/api/eventos/evento_activo_en_canal/{id}": {
            "post": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Obtener evento activo dado un canal",
                "description": "Obtener evento activo dado un canal",
                "operationId": "App\\Http\\Controllers\\EventosController::eventoActivoEnCanal",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de canal",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Canal obtenido con éxito"
                    },
                    "404": {
                        "description": "Evento no encontrado"
                    }
                }
            }
        },
        "/api/eventos/{id}": {
            "put": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Actualizar un evento",
                "description": "Actualizar un evento",
                "operationId": "App\\Http\\Controllers\\EventosController::update",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Título del evento",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Descripción del evento",
                                        "type": "string"
                                    },
                                    "inicio": {
                                        "description": "Fecha/Hora de Inicio",
                                        "type": "DATETIME"
                                    },
                                    "final": {
                                        "description": "Fecha/Hora de el Final",
                                        "type": "DATETIME"
                                    },
                                    "delegacion_id": {
                                        "description": "Id de la Delegación",
                                        "type": "integer"
                                    },
                                    "views": {
                                        "description": "Views",
                                        "type": "integer"
                                    },
                                    "likes": {
                                        "description": "Likes",
                                        "type": "integer"
                                    },
                                    "max_online": {
                                        "description": "Max Online",
                                        "type": "integer"
                                    },
                                    "canal_id": {
                                        "description": "Id de Canal",
                                        "type": "integer"
                                    },
                                    "estado": {
                                        "description": "estado de el evento ['EN_ESPERA','EMITIENDO','FINALIZADO']",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Evento no encontrado"
                    }
                }
            }
        },
        "/api/eventos/ver/{id}": {
            "post": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Ver un evento como usuario",
                "description": "Ver un evento como usuario",
                "operationId": "App\\Http\\Controllers\\EventosController::VerEvento",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de evento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Evento no encontrado"
                    }
                }
            }
        },
...
    }
}